+++ source/scenario/scenario.h
@@ -132,7 +132,11 @@ short scenario_get_fog_region_index(const struct location *location, const real_
 boolean scenario_location_underwater(const struct location *location, const real_point3d *position, short *optional_weather_palette_index);
 real scenario_location_water_depth(const struct location *location, const real_point3d *position);
 void scenario_get_sound_environment(long *background_sound_index, long *sound_environment_tag, boolean *crossed_water_boundary);
-void scenario_get_atmospheric_fog(short local_player_index, long sky_index, real_point3d *camera_point, struct render_fog *render_fog);
+void scenario_get_atmospheric_fog(
+	short local_player_index,
+	short sky_index,
+	real_point3d *camera_point,
+	struct render_fog *render_fog);
 
 boolean scenario_get_current(
 	struct location const *location,
